[feat]: dashboard search modal
